Add missing methods

+++ Rust-MCP-Server/src/request_stats.rs
@@ -1,6 +1,6 @@
 #[derive(Debug, Clone)]
 pub struct RequestStats {
     pub todo_count: usize,
-    pub file_count: u32,
-    pub unfinished_tasks: u32,
+    pub file_count: usize,
+    pub unfinished_tasks: usize,
 }
